EFA_0015_FEATURE_LATEENTRY_SHALL_PRESENT_LAST_SESSION
Feature for late entry:
If the user did neither choose a boat nor a person when using the "late entry" function, and if the latest logbook record has been entered just 2 minutes earlier or less, the data of the latest logbook entry is presented.

This helps if you have to provide multiple late entries for a group of persons who all were on the same trip.