EFA_0016_EFA_LOG_FILE_INCLUDED_IN_BACKUP
When a backup is created, the CURRENT logfile log/efa.log is now included within the zip file.

EFA2 provides some log rotation during startup, when the current logfile exceeds 1M. These rotated log files are not included in the backup.

Also, the efa log file is not included in the restore function of efa. This is an intended behaviour: when you restore a backup, it is very likely you want to KEEP your current log file instead of replacing it by an old one.