Less config items, config items are TYPE_EXPERT
Reduced config values for filter textfields above lists from 4 to 2:

- Common lists (available boats, boats on the water, persons)
- not available boats
The latter is because filtering in this does not make sense for smaller clubs and the text field wastes vertical screen size on smaller screens.

Config items are TYPE_EXPERT

Moved the new config items to an earlier position in the "Boathouse" category (they are more likely to be searched for by the admin at that position).

Updated international support and removed no longer needed elements of this feature.