Quick entry of new persons detects association name
See https://github.com/nicmichael/efa/issues/23

When starting a new session, and enter a new person's name, the new syntax is
So the correct pattern for entering such a name should be

    Lastname, Firstname (Nameaddon) [NameOfAssociation]
    respectively
    Firstname Lastname (Nameaddon) [NameOfAssociation]

so that the name of the association is detected.
Also, the quick entry dialog contains the field for the Association, so that it can be changed (e.g. because of typos) before saving the entry.