Bugfix for filtered lists in efaBoatHouse main screen
ItemTypeList.java
--------------------------
setupPanel() - if a list has a visible filter textfield, this filter text field should get the focus if the list itself gets the focus.
When efaBoatHouse is idle for more than one minute, the background task will set the focus on the list of the available boats. The effect of this line is, that the filter field gets the focus, so that the user can enter the search criteria without focussing the filterfield by using the mouse.

filteredSize()
This function returns the number of items which are shown in the list, instead of the number of items which are in the datalist behind the gui element.

EfaBoathouseFrame.java
-----------------------------
showBoatStatus -
Bugfix: an ItemList does not always have a selected item. So the name is only set when getSelectedListItem is not null.

Bugfix: loop must end, not when the itemindex exceeds the list's data size, but when it exceeds the number of SHOWN elements within the list. If it is filtered, there is a difference between data size and model size.