MN_EFA_027_FIX_EFACLI_LOGROTATE_PROBLEM
Summary
---------
See my notes in this thread of the efa support forum:
http://forum.nmichael.de/viewtopic.php?f=15&t=1214&p=4792#p4792

Only efaBths and efaBase will perform a logrotate on startup. Other efa programs will simply log into an existing logfile until it reaches 100MB.

The file size when efabootshaus performs a logrotate is set to 10MB.
All efa programs will quit logging when the file size exceeds 100MB.

Details:
---------
So in conclusion, a rotation of the logfile only takes place during the *startup* of efaBase or efaBths.
This usually takes place at 4am, or if the programs get started manually.
efaBase and efaBths cannot be run at the same time on the same machine, so there is no problem that
two instances are trying to rotate the log file simultaneously.

efaCLI itself does need a running efa_boathouse oder efa_base in background to fulfill its task.
So, during the startup of efaBths or efaBase, there no efaCLI program can log something
as efaCLI cannot connect to efaBths or efaBase. Even if these efaCLI log entries are lost (as they are written
to a log file no longer existing), it would not matter much - during startup, efaCLI only logs technical informations.

This fix is not final, but it pretty much solves the problem stated in the forum (http://forum.nmichael.de/viewtopic.php?f=15&t=1214&p=4792#p4792):
There are no two concurrent efa programs which are trying to rotate the log simultaneously.

A final fix would be switching to log4J and configuring it to produdent mode.