MN_EFA_041_BOATRESERVATION_COPY_AND_CHANGE_BOAT
Summary
--------------
Base for this feature is a request in efa2 forum: http://forum.nmichael.de/viewtopic.php?f=9&t=1154&p=4636

Within the boat reservation dialog there are two new functions
- Create copy
- Change boat (for current reservation)

The workflow is the same as it is for creating a new reservation.
First the user has to specify the (new) boat name, and second the user can change some other entries of the reservation record.

This workflow is established due to the fact that for boat reservations, the boat name is part of the primary key of a reservation.
So, the boat cannot be changed after saving the reservation initially.

ItemTypeDataRecordTable.java
------------
Small refactoring due to adding some documentation on the ACTION IDs of the buttons.

BoatReservations.java
------------
Small refactoring: put checking for overlapping reservations in a new method

DataTypeDate.java
------------
Bugfix: isRangeOverlap() in the variant with separate date and time fields differed from the date-based variant.

This bug appeared only if you created a 1:1 copy of a boat reservation where boat AND date and time fields are exactly the same. This is a standard case for creating a copy of a boat reservation, so this bug has been fixed.

BaseDialog.Java, ImagesAndIcons.Java
------------
Refactoring: All Icons now are within a single class and referenced by a constant instead of filenames.
So it is easier to find out which icon is used by tracing the constants.

Also refactored icons in
DataPrintRecordDialog
EfaBoathouseFrame
MultiInputDialog
SimpleInputDialog
SimpleOptionInputDialog
FixLogbookDialog
StatisticsEditDialog
UnversionizedDataEditDialog
EfaMenuButton

BoatReservationListDialog.java
-----------
Added some documentation how and from which sources BoatReservationListDialog gets the buttons for the actions.

DataListDialog.java
-----------
As DataListDialog gets its buttons from several sources, and we only want do ADD Buttons for BoatReservationListDialog, it must be possible to add a button at a certain position within the existing button array.

So this is what the new Method insertAction() is for.

efa_de, efa_en, efa_fr.properties
------------
added new translations