Bugfix for failing data import of Logbook items
Summary
------------
Bug description: https://github.com/nicmichael/efa/issues/105

Data import of logbook items did not work at all.

Fixed determination of field type for fields (in a CSV file) that do not have an exact equivalent in a DataRecord.

DataImport.java
-------
- Added some comment about Datarecord.getFieldType() behaviour.

DataRecord.java
-------
- int getFieldType(fieldName) Method
Added method documentation regarding behaviour when a given fieldName does not exist in Metadata

MetaData.java
-----------
- int getFieldType(fieldName) method:
If the fieldIndex of the given fieldName cannot be determined, it is -1. Instead of determining fieldType of a nonexistant field, we return IDataAccess.DATA_UNKNOWN.

This does not hurt for the existing code, and is better than throwing an Exception here.

IDataAcess.java
------------
- added public constant for a fieldtype DATA_UNKNOWN for nonexistant fields.

This