Bugfix in AutoCompleteList
Bug#1 a TAB key (or a focuslost event) on a autocomplete element always took over the actual selected element. If the popup list opened by accident, this lead to unwanted commentment to the selected value.

So, when hitting the TAB key or loosing focus by klicking somewhere else, the selected element is only taken if the user entered at least one character in the text field.

Bug#2 AutoComplete field not showing
This happens under windows only. If you open the lookup window on a field, then click into another program and then back into efaBths in the currently open field, the lookup list never showed up on windows. It stayed in the background.

By now when returning into the mask with the open autocomplete window, you just have to hit a key so that the popup window shows again.