Small bugfix, Club logo up to 320x200
Summary
------
The Logbook dialog which can be run from efaBths main screen did not look good if there were a lot of trips with more than two crew members. So alternating color for the crew table within a row has been deactivated.

Also, current monitors provide a lot of space when they are 16:10 or 16:9. The logo can consume more space and thus be bigger.
So the logo can be up to 320x200 pixels.

The logo width affects the widths of the buttons in efaBths. So the wider the logo, the wider the widths of the buttons.
This is intended; the user has to take care by himself that the look of efaBths is not affected badly by too big logos.

The FileOpenDialog now supports up to three file extensions. PNG file format has been added as it is supported by java 1.8+

StatisticsDialog now uses more pixels for the "number" column.