Initial Version
Summary
---------
For the user, the sortorder of boat reservations is more convenient.

The user may choose any column to sort the boat reservations. No change here.

There is a secondary sortorder by the amount of time needed to pass so that the reservation gets active.

When ordering by DESCRIPTION (by clicking on the corresponding column header), the table is ordered by DESCRIPTION ASC, MINUTES_UNTIL_RESERVATION_GETS_ACTIVE ASC, BOATNAME ASC.

So the sort order of the reservations always is in the order the reservations get activated.