Update Waters.java
Summary
-----------
Efa determines automatically if a new waters file is present (in its JAR). If yes, it asks to update the waters once admin mode is opened.

When running a beta belease (2.4.1_#1), the beta is marked by an # character.

This conflicts with the automatic recognition of efa for new waters files, as the # is used there for splitting last version number and a water's MD5 hash.

The water's MD5 hash get a different delimiter (|) so that beta versions of efa can use the former template (2.4.1_#1).