MN_EFA_032_KANUEFB_SYNC_
Summary
-----------
When using Kanu-EFB synchronisation (available for german canoers only),
you need to enter the Kanu-EFB-Id for any person you want to enable the
synchronisation.

As the Kanu-EFB-ID fields are in expert mode, you need to enable the expert mode
any time you want to enter a Kanu-EFB-ID.

This is quite enerving. So there is now an option within the "Sync" tab
to tell efa it shall always show these Kanu-EFB-ID fields.

This applies for Persons, Boats, Waters, although kanu-efb sync nowadays only needs
a Kanu-EFB-ID for Persons.

On the other hand, the Kanu-EFB-ID can also be filled with the Kanu-EFB user name
of the club member. This is a new feature of Kanu-EFB as of 02/2024, and not that obvious.
So a hint is shown in the edit person dialog right above the Kanu-EFB-ID field.