BugFix when names of non-oneseater boats were put in.
If a user entered the name of a non-oneseater boat, efa would treat it as a foreign boat.

unfortunately, the audit task later would correct the link to the "real" non-one-seater boat.

So this would look bad in the logbook, if a single person would try to manage a four-seater boat.

So efa now checks for the boat names with an orange or red box behind the input field, if they match to a boat which has NO one-seater configuration.

Then the multisession dialog cannot be closed.