Fixed nullpointer, handled DataTypeDistance
No more Nullpointer exceptions when exporting records in Debug mode, if a field is a virtual field.

the efaLogbook record has a lot of virtual fields, like for Crew1...Crew24.

Also, the double part of distance fields get formatted in a CSV locale, if a locale is specified.

Added better documentation.