CSV with BOM choosable by user again
In the prior version, efa determined wether or not to create a BOM in the CSV file by checking the charset for UTF-8.

As some users in the forums pointed out, PHP programs are not very likely to support CSV files with a BOM, no matter wether they use UTF-8 character set.

So, efa provides the user with the option to create a csv file with or without a BOM.

Also, a small hint is shown above the file format radio button group.

User can also determine CSV format in efaCLI.